fix(review): address PR review findings — unused imports, stub dedup, test cohesion
- Remove unused asyncio/AsyncMock imports from dispatch_params and dispatch_validation
- Consolidate _simple_prompt_builder/_no_sleep_quota_checker/_noop_quota_refresher into
  tests/server/_helpers.py; remove verbatim copies from dispatch, dispatch_params, dispatch_validation
- Move test_dispatch_food_truck_marketplace_install_succeeds to test_tools_dispatch.py
  (execution lifecycle test, not parameter passthrough)
- Move test_checks_fetch_result to TestMergeWorktreeRemoteTrackingGuard in
  test_tools_git_merge_guards.py (fetch failure is a pre-rebase guard, not cleanup reporting)
- Add test_tools_kitchen_gate_features.py to _SERVER_FILE_LEVEL_ENTRIES in
  test_test_filter_cascade.py (aligns with _test_filter.py line 596)
- Extract _HOOK_CONFIG_OVERLAY_RELPATH to _helpers.py; import in hook_config test
- Fix CLAUDE.md description for test_tools_dispatch_params.py (removed "marketplace")
- Remove stale [FAILS NOW]/[NEW COVERAGE] docstring tags from test_tools_git_classify_fix.py
- Remove REQ-PACK-008, T-REFRESH-*, Group K label comments from kitchen_gate_features.py
