+++ tests/server/test_tools_dispatch.py
@@ -11,29 +11,15 @@ import pytest
 
 from autoskillit.fleet import FleetSemaphore
 from tests.fakes import InMemoryHeadlessExecutor, InMemoryRecipeRepository
-from tests.server._helpers import _make_recipe_info, _make_standard_recipe
+from tests.server._helpers import (
+    _make_recipe_info,
+    _make_standard_recipe,
+    _no_sleep_quota_checker,
+    _noop_quota_refresher,
+    _simple_prompt_builder,
+)
 
 pytestmark = [pytest.mark.layer("server"), pytest.mark.medium, pytest.mark.feature("fleet")]
-
-
-def _simple_prompt_builder(**kwargs) -> str:
-    """Minimal prompt builder for tests — avoids CLI imports."""
-    return f"prompt-for-{kwargs.get('recipe', 'unknown')}"
-
-
-async def _no_sleep_quota_checker(config, **kwargs) -> dict:
-    """Quota checker stub: always returns no-sleep result."""
-    return {
-        "should_sleep": False,
-        "sleep_seconds": 0,
-        "utilization": None,
-        "resets_at": None,
-        "window_name": None,
-    }
-
-
-async def _noop_quota_refresher(config, **kwargs) -> None:
-    """Quota refresher stub: no-op."""
 
 
 class TestDispatchFoodTruckExecution:
@@ -370,4 +356,45 @@ class TestDispatchFoodTruckExecution:
         )
 
         result = json.loads(result_json.outcome.to_envelope())
+        assert result["success"] is True
+
+    @pytest.mark.anyio
+    async def test_dispatch_food_truck_marketplace_install_succeeds(
+        self, tool_ctx_marketplace, monkeypatch
+    ):
+        """dispatch_food_truck does not raise when plugin_source is MarketplaceInstall."""
+        from autoskillit.fleet._api import execute_dispatch
+        from autoskillit.fleet.result_parser import L3ParseResult
+
+        monkeypatch.setattr(
+            "autoskillit.fleet._api.parse_l3_result_block",
+            lambda **_kwargs: L3ParseResult(
+                outcome="completed_clean",
+                payload={"success": True},
+                raw_body=None,
+                parse_error=None,
+                source="stdout",
+            ),
+        )
+
+        tool_ctx_marketplace.fleet_lock = FleetSemaphore(max_concurrent=1)
+        repo = InMemoryRecipeRepository()
+        recipe_info = _make_recipe_info("test-recipe")
+        repo.add_recipe("test-recipe", recipe_info)
+        repo.add_full_recipe(recipe_info.path, _make_standard_recipe("test-recipe", ["task"]))
+        tool_ctx_marketplace.recipes = repo
+        tool_ctx_marketplace.executor = InMemoryHeadlessExecutor()
+
+        _dispatch_result = await execute_dispatch(
+            tool_ctx=tool_ctx_marketplace,
+            recipe="test-recipe",
+            task="t",
+            ingredients=None,
+            dispatch_name=None,
+            timeout_sec=None,
+            prompt_builder=_simple_prompt_builder,
+            quota_checker=_no_sleep_quota_checker,
+            quota_refresher=_noop_quota_refresher,
+        )
+        result = json.loads(_dispatch_result.outcome.to_envelope())
         assert result["success"] is True
